feat: rearrange the files to comply with sveltekit v2 standards
