+++ src/routes/+layout.server.ts
@@ -1,0 +1,12 @@
+import type { LayoutServerLoad } from './$types';
+import getVisitors from '$lib/server/counter';
+
+export const load: LayoutServerLoad = async () => {
+  const visitorCount = await getVisitors();
+
+  return {
+    count: {
+      count: visitorCount
+    }
+  };
+};
