--- src/routes/+layout.ts
@@ -1,2 +1,0 @@
-export const csr = true;
-export const ssr = false;
